backend/requirements.txt
fastapi
uvicorn
nylas
python-dotenv
sqlalchemy
psycopg2-binary
temporalio
fastapi-cors